backend/requirements.txt
opencv-python==4.9.0.80
tensorflow==2.15.0.post1
pandas==2.2.1
numpy==1.26.4
flask==2.3.2  # Compatible Flask version
flask-cors==3.0.10  # Added to handle CORS
Pillow==10.0.0  # Added to handle images
scikit-learn==1.3.1  # Compatible version of scikit-learn
apscheduler